+++ lib/src/features/menu/models/Section.dart
@@ -1,0 +1,11 @@
+import 'Drink.dart';
+
+class Section {
+  final String title;
+  final List<Drink> drinks;
+
+  Section({
+    required this.title,
+    required this.drinks,
+  });
+}
