+++ lib/src/theme/app_colors.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+sealed class AppColors {
+  static const blue = Color(0xC20F86BD);
+  static const white = Color(0xFFE7F0F7);
+  static const black = Colors.black;
+  static const milk = Color(0xFFF7FAF8);
+}
